RED: Extend the bad fact property filter
**What this does**
Make the bad match property filter on a regex that expresses both order
of brackets and having both brackets as a requirement for goodness. If
that good regex isn't matched, then this is a bad fact.

**Why this is being done**
The original property was too weak - "does not have [" was not enough,
as is demonstrated by the first failure under the new tests - a single
"[" string is passing as a good match, when clearly it is not.