YELLOW: TestForFail per module
**What this does**
Resolves https://github.com/toejough/spacer/issues/23

**Why this is being done**
Because rapid test flags were causing test failures now that not every
module under the root repo level use rapid flags. Now that the modules
are tested independently, the rapid flags are valid for the modules that
import `rapid`, and are not used (and thus not erroring) for the ones
that do not.