DEV: Make mock functions come from the fut
**What this does**
Makes mocks come from the function under test so that the shared call
channel is just handled internally and the user doesn't have to deal
with it.

Also updates names to be clearer about what mocks are and are not
capable of wrt args and returns.

Also replaces the hard-coded "true" with a variable that preserves
meaning.

**Why this is being done**
To streamline the test to only make the user think about what they need
to in terms of testing and expectations.