DEV: Simplify Eq generic use
**What this does**
Replaces [S any, A Eq[S]] with [A Eq[A]] because that's simpler and more
specifically what we want. It also allows the compiler to do the type
checking on the .Equals method calls better.

Also updates the functions with methods for the ExpectCall callables,
because we can reliably use the generic types on the receiver for the
functions/methods.

**Why this is being done**
to simplify the signatures, make the code more robust / compile-time
safe, and make the code easier to read.