DEV: Restrict necessity for eq type
**What this does**
Restrict the necessity for the eq type to only where it's actually used.
Everywhere else, allow any type for the args.

**Why this is being done**
To allow the end user to use any type they want for args, without having
to create the equality type (if they don't want to use the fatal helper
method)