refactor(test): Use concrete dependency structs
**What this does**
Use concrete dependency structs instead of interfaces.

**Why this is being done**
Reduce levels of indirection, which increases clarity of functionality.
Reduce code, which reduces the number of places bugs can hid.
Allow dependency functions to be generic, which they could not be as
methods of a struct that met an interface's requirements.

test-status: RED