poc: prompted to make the buttons the same style
give the delete and reveal buttons the same style

Then I had to prompt it to put the delete button back in the right
place, because it'd moved it...

put the delete button on the right side of the question

This was the most trouble I've had with the prompting. I tried 3 or 4
variations before landing on this one.